Hoist shared types and interfaces into zephyr-grader-base
