+++ packages/zephyr-catch-grader/package.json
@@ -23,6 +23,7 @@
   "license": "NCSA",
   "dependencies": {
     "@babel/runtime": "^7.0.0-beta.54",
+    "@illinois/zephyr-grader-base": "^1.0.0",
     "babel-polyfill": "^6.26.0",
     "deasync": "^0.1.13",
     "rxjs": "^6.2.2",
@@ -38,12 +39,15 @@
     "@babel/preset-env": "^7.0.0-beta.54",
     "@babel/preset-typescript": "^7.0.0-beta.54",
     "@types/deasync": "^0.1.0",
+    "@types/fs-extra": "^5.0.4",
     "@types/jest": "^23.3.0",
     "@types/node": "^10.5.2",
     "@types/validator": "^9.4.1",
     "@types/xml2js": "^0.4.3",
     "jest": "^23.4.1",
     "jest-plugin-fs": "^2.9.0",
+    "merge-trees": "^2.0.0",
+    "tmp-promise": "^1.0.5",
     "ts-jest": "^23.0.1",
     "tslint": "^5.11.0",
     "typescript": "^2.9.2"
