+++ packages/zephyr-grader-base/.babelrc
@@ -1,0 +1,11 @@
+{
+  "presets": [
+      "@babel/env",
+      "@babel/typescript"
+  ],
+  "plugins": [
+      "@babel/plugin-transform-runtime",
+      "@babel/proposal-class-properties",
+      "@babel/proposal-object-rest-spread"
+  ]
+}
