+++ packages/zephyr-grader-base/.npmignore
@@ -1,0 +1,3 @@
+src
+.babelrc
+tsconfig.json
