+++ packages/zephyr-grader-base/tsconfig.json
@@ -55,8 +55,5 @@
     /* Experimental Options */
     // "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
     // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
-  },
-  "include": [
-    "src/**/*"
-  ]
+  }
 }
