+++ packages/zephyr-staff-cli/package.json
@@ -27,6 +27,7 @@
     "@babel/runtime": "^7.0.0-beta.54",
     "@illinois/zephyr-catch-grader": "^1.0.3",
     "@illinois/zephyr-github-checkout": "^1.0.3",
+    "@illinois/zephyr-grader-base": "^1.0.0",
     "@octokit/rest": "^15.9.4",
     "babel-polyfill": "^6.26.0",
     "csv-stringify": "^3.1.1",
